# infrastructure/restapi/controllers/book/responses.go
// Package book contains the book controller
package book

// MessageResponse is a struct that contains the response body for the message
type MessageResponse struct {
	Message string `json:"message"`
}